Move getPreviousMove to a different file
Next commit will make getPreviousMove be used outside of just netscript, so it didn't belong in the netscript-specific file anymore
